requirements.txt
# aiohttp==3.9.5
# aiosignal==1.3.1
# altair==5.3.0
# annotated-types==0.7.0
# anyio==4.4.0
# arxiv==2.1.0
# asgiref==3.8.1
# asttokens==2.4.1
# attrs==23.2.0
# backoff==2.2.1
# bcrypt==4.1.3
# beautifulsoup4==4.12.3
# blinker==1.8.2
# bs4==0.0.2
# build==1.2.1
# cachetools==5.3.3
# cassandra-driver==3.29.1
# cassio==0.1.8
# certifi==2024.6.2
# charset-normalizer==3.3.2
# chroma-hnswlib==0.7.3
# click==8.1.7
# colorama==0.4.6
# coloredlogs==15.0.1
# comm==0.2.2
# dataclasses-json==0.6.7
# debugpy==1.8.1
# decorator==5.1.1
# Deprecated==1.2.14
# distro==1.9.0
# dnspython==2.6.1
# email_validator==2.1.2
# executing==2.0.1
# faiss-cpu==1.8.0
# fastapi==0.115.6
# fastapi-cli==0.0.4
# feedparser==6.0.10
# filelock==3.15.3
# flatbuffers==24.3.25
# frozenlist==1.4.1
# fsspec==2024.6.0
# geomet==0.2.1.post1
# gitdb==4.0.11
# GitPython==3.1.43
# google-auth==2.30.0
# googleapis-common-protos==1.63.1
# greenlet==3.0.3
# groq==0.9.0
# grpcio==1.68.0
# grpcio-tools==1.62.3
# h11==0.14.0
# h2==4.1.0
# hpack==4.0.0
# httpcore==1.0.5
# httptools==0.6.1
# httpx==0.27.0
# huggingface-hub==0.23.4
# humanfriendly==10.0
# hyperframe==6.0.1
# idna==3.7
# importlib_metadata==7.1.0
# importlib_resources==6.4.0
# intel-openmp==2021.4.0
# ipykernel==6.29.4
# ipython==8.25.0
# jedi==0.19.1
# Jinja2==3.1.4
# joblib==1.4.2
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonschema==4.22.0
# jsonschema-specifications==2023.12.1
# jupyter_client==8.6.2
# jupyter_core==5.7.2
# kubernetes==30.1.0
# langchain==0.3.10
# langchain-chroma==0.1.4
# langchain-community==0.3.10
# langchain-core==0.2.12
# langchain-groq==0.2.1
# langchain-openai==0.2.11
# langchain-text-splitters==0.2.1
# langchainhub==0.1.20
# langserve==0.2.2
# langsmith==0.1.80
# lark==1.2.2
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# marshmallow==3.21.3
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# mkl==2021.4.0
# mmh3==4.1.0
# monotonic==1.6
# mpmath==1.3.0
# multidict==6.0.5
# mypy-extensions==1.0.0
# nest-asyncio==1.6.0
# networkx==3.3
# numpy==1.26.4
# oauthlib==3.2.2
# objectbox==4.0.0
# onnxruntime==1.18.0
# openai==1.57.0
# opentelemetry-api==1.25.0
# opentelemetry-exporter-otlp-proto-common==1.25.0
# opentelemetry-exporter-otlp-proto-grpc==1.25.0
# opentelemetry-instrumentation==0.46b0
# opentelemetry-instrumentation-asgi==0.46b0
# opentelemetry-instrumentation-fastapi==0.46b0
# opentelemetry-proto==1.25.0
# opentelemetry-sdk==1.25.0
# opentelemetry-semantic-conventions==0.46b0
# opentelemetry-util-http==0.46b0
# orjson==3.10.5
# overrides==7.7.0
# packaging==23.2
# pandas==2.2.2
# parso==0.8.4
# pillow==10.3.0
# platformdirs==4.2.2
# portalocker==2.10.0
# posthog==3.5.0
# prompt_toolkit==3.0.47
# protobuf==4.25.3
# psutil==6.0.0
# pure-eval==0.2.2
# pyarrow==16.1.0
# pyasn1==0.6.0
# pyasn1_modules==0.4.0
# pydantic==2.10.3
# pydantic_core==2.18.4
# pydeck==0.9.1
# Pygments==2.18.0
# pypdf==4.2.0
# PyPDF2==3.0.1
# PyPika==0.48.9
# pyproject-toml==0.0.10
# pyproject_hooks==1.1.0
# pyreadline3==3.4.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# python-multipart==0.0.9
# pytz==2024.2
# pywin32==308
# PyYAML==6.0.1
# pyzmq==26.0.3
# qdrant-client==1.10.0
# referencing==0.35.1
# regex==2024.5.15
# requests==2.31.0
# requests-oauthlib==2.0.0
# rich==13.7.1
# rpds-py==0.18.1
# rsa==4.9
# safetensors==0.4.3
# scikit-learn==1.5.0
# scipy==1.13.1
# sentence-transformers==3.0.1
# sgmllib3k==1.0.0
# shellingham==1.5.4
# six==1.16.0
# smmap==5.0.1
# sniffio==1.3.1
# soupsieve==2.5
# SQLAlchemy==2.0.31
# sse-starlette==2.1.2
# stack-data==0.6.3
# starlette==0.37.2
# streamlit==1.35.0
# sympy==1.12.1
# tbb==2021.12.0
# tenacity==8.4.1
# threadpoolctl==3.5.0
# tiktoken==0.7.0
# tokenizers==0.19.1
# toml==0.10.2
# toolz==0.12.1
# torch==2.3.1
# tornado==6.4.1
# tqdm==4.66.4
# traitlets==5.14.3
# transformers==4.41.2
# typer==0.12.3
# types-requests==2.32.0.20240602
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# tzdata==2024.1
# ujson==5.10.0
# urllib3==2.2.2
# uvicorn==0.32.1
# watchdog==4.0.1
# watchfiles==0.22.0
# wcwidth==0.2.13
# websocket-client==1.8.0
# websockets==12.0
# wikipedia==1.4.0
# wrapt==1.16.0
# yarl==1.9.4
# zipp==3.19.2

#These are the packages needed for our chatbot
fastapi
langchain-groq
langchain
langchain-openai
langchain-chroma
python-dotenv
langchain-community
scikit-learn
lark





